requirements.txt.txt
astroid==2.2.5
autopep8==1.4.4
colorama==0.4.1
Django==2.2.5
django-upload-validator==1.0.2
isort==4.3.21
lazy-object-proxy==1.4.2
libmagic==1.0
mccabe==0.6.1
pep8==1.7.1
Pillow==6.1.0
psycopg2==2.8.3
pycodestyle==2.5.0
pylint==2.3.1
python-magic==0.4.15
python-magic-bin==0.4.14
pytz==2019.2
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2